Login/logout and routing in Vue.js
Made it possible to log in via Vue.js and also log out. Log out will clear the session so no user can log in again without entering their credentials again. Also depending on the user role (admin or user) it will redirect you to the user page (if user) or admin page (if admin). Made it impossible for a User to access the Admin page. Only the Admin page can log into the Admin page. Cleaned the code a bit and corrected some spacing and comments