100% completition of Login/Register and Routing
Finished the log in and register pages. The routes '/user' and '/admin' cannot be accessed unless you are logged in and have the corresponding account role (ex: To access '/admin' your account needs to have 'role: admin'). Created a LoginBar component in Vue, that contains the Log in button, Log out, and Link to corresponding User/admin page (depending on who is logged in. Made it so that the User or Admin page will show the account's name when inside of it. Cleaned the code a lot, formatted the code, styled several pages(see file changes).